Refactor: 리퀘스트, 리스폰스 통일
2021_01_03
