--- Portfolian/app/src/main/java/com/example/portfolian/data/KakaoToken.kt
@@ -1,5 +1,0 @@
-package com.example.portfolian.data
-
-data class KakaoToken (val token: String)
-
-data class RenewalTokenRequest(val refreshToken: String, val userId: String)
