--- Portfolian/app/src/main/java/com/example/portfolian/data/NickName.kt
@@ -1,3 +1,0 @@
-package com.example.portfolian.data
-
-data class NickName(val nickName: String)
