--- Portfolian/app/src/main/java/com/example/portfolian/data/NickNameResponse.kt
@@ -1,10 +1,0 @@
-package com.example.portfolian.data
-
-import com.google.gson.annotations.SerializedName
-
-class NickNameResponse (
-    @SerializedName("code")
-    val code: Int,
-    @SerializedName("message")
-    val message: String
-    )
