--- Portfolian/app/src/main/java/com/example/portfolian/data/OauthResponse.kt
@@ -1,17 +1,0 @@
-package com.example.portfolian.data
-
-import com.google.gson.annotations.SerializedName
-
-
-data class OauthResponse(
-    @SerializedName("code")
-    val code: Int,
-    @SerializedName("isNew")
-    val isNew: Boolean,
-    @SerializedName("refreshToken")
-    val refreshToken: String,
-    @SerializedName("accessToken")
-    val accessToken: String,
-    @SerializedName("userId")
-    val userId: String
-)
