--- Portfolian/app/src/main/java/com/example/portfolian/data/ReadProjectResponse.kt
@@ -1,49 +1,0 @@
-package com.example.portfolian.data
-
-import android.os.Parcelable
-import com.google.gson.annotations.SerializedName
-import kotlinx.android.parcel.Parcelize
-
-data class ReadProjectResponse(
-    @SerializedName("code")
-    var code: Int,
-
-    @SerializedName("articleList")
-    var articleList: ArrayList<Project>
-)
-
-@Parcelize
-data class Project(
-    @SerializedName("projectId")
-    var projectId: String,
-
-    @SerializedName("title")
-    var title: String,
-
-    @SerializedName("stackList")
-    var stackList: List<String>,
-
-    @SerializedName("description")
-    var description: String,
-
-    @SerializedName("capacity")
-    var capacity: Int,
-
-    @SerializedName("view")
-    var view: Int,
-
-    @SerializedName("bookMark")
-    var bookMark: Boolean,
-
-    @SerializedName("status")
-    var status: Int,
-
-    @SerializedName("leader")
-    var leader: Leader
-) : Parcelable
-
-@Parcelize
-data class Leader(
-    var userId: String,
-    var photo: String
-) : Parcelable
