--- Portfolian/app/src/main/java/com/example/portfolian/data/TokenResponse.kt
@@ -1,12 +1,0 @@
-package com.example.portfolian.data
-
-import com.google.gson.annotations.SerializedName
-
-data class TokenResponse (
-    @SerializedName("code")
-    val code: Int,
-    @SerializedName("accessToken")
-    val accessToken: String,
-    @SerializedName("message")
-    val message: String
-    )
