--- Portfolian/app/src/main/java/com/example/portfolian/data/WriteProject.kt
@@ -1,6 +1,0 @@
-package com.example.portfolian.data
-
-data class WriteProject (
-    var userId: String,
-    var article: Article,
-    var ownerStack: String)
